pyproject.toml
[project]
name = "telegram-group-message-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "telethon>=1.40.0",
]
